+++ CleanestHud/HudChanges/HudDetails.cs
@@ -17,14 +17,22 @@ namespace CleanestHud.HudChanges
     {
         internal static class AssetEdits
         {
-            internal static void RemoveHudElementPrefabDetails()
+            internal static void EditHudElementPrefabDetails()
             {
+                EditItemIconNavHighlight();
                 RemoveMonsterPanelsHudDetails();
                 RemoveGameEndPanelDetails();
                 RemoveScoreboardStripAssetDetails();
                 RemoveMoonDetonationPanelDetails();
                 RemoveStatStripTemplateImage();
                 RemoveChatBoxDetails();
+            }
+            private static void EditItemIconNavHighlight()
+            {
+                Transform navFocusHighlightTransform = HudAssets.ItemIconPrefab.transform.GetChild(1);
+
+                RawImage navFocusHighlightRawImage = navFocusHighlightTransform.GetComponent<RawImage>();
+                navFocusHighlightRawImage.texture = ModAssets.AssetBundle.LoadAsset<Texture2D>("NewNavHighlight");
             }
             private static void RemoveMonsterPanelsHudDetails()
             {
@@ -45,6 +53,8 @@ namespace CleanestHud.HudChanges
                 Image inventoryDisplayImage = inventoryDisplay.GetComponent<Image>();
                 inventoryDisplayImage.enabled = false;
             }
+
+
             // TODO i think the chatbox here doesn't get edited? idk look into it
             // also finish this
             private static void RemoveGameEndPanelDetails()
@@ -147,6 +157,9 @@ namespace CleanestHud.HudChanges
             }
             private static void RemoveScoreboardStripAssetDetails()
             {
+                RawImage scoreboardStripRawImageBackground = HudAssets.ScoreboardStrip.gameObject.GetComponent<RawImage>();
+                scoreboardStripRawImageBackground.texture = ModAssets.AssetBundle.LoadAsset<Texture2D>("NewNavHighlight");
+
                 Transform longBackground = HudAssets.ScoreboardStrip.transform.Find("LongBackground");
 
                 Transform classBackground = longBackground.Find("ClassBackground");
@@ -485,6 +498,37 @@ namespace CleanestHud.HudChanges
             equipmentTextBackgroundPanelImage.enabled = false;
         }
 
+
+
+        internal static void SetupItemIconGlowImageForColoring(ItemIcon itemIcon)
+        {
+            itemIcon.glowImage = itemIcon.transform.GetChild(1).GetComponent<RawImage>();
+            HGButton button = itemIcon.gameObject.GetComponent<HGButton>();
+            // setting the button colors to white makes it not influence glowimage color with yellow (the default)
+            // we can always just manually color it back to yellow later anyways
+            button.m_Colors.highlightedColor = Main.Helpers.ChangeColorWhileKeepingAlpha(button.m_Colors.highlightedColor, Color.white);
+            button.m_Colors.normalColor = Main.Helpers.ChangeColorWhileKeepingAlpha(button.m_Colors.normalColor, Color.white);
+            button.m_Colors.pressedColor = Main.Helpers.ChangeColorWhileKeepingAlpha(button.m_Colors.pressedColor, Color.white);
+            button.m_Colors.selectedColor = Main.Helpers.ChangeColorWhileKeepingAlpha(button.m_Colors.selectedColor, Color.white);
+            // better higlight visibility for some character colors
+            button.m_Colors.m_ColorMultiplier = 1.5f;
+        }
+
+        internal static void EditScoreboardStripEquipmentSlotHighlight(ScoreboardStrip scoreboardStrip)
+        {
+            Transform equipmentBackground = scoreboardStrip.equipmentIcon.transform;
+            Transform navFocusHighlight = equipmentBackground.GetChild(1);
+
+            RawImage equipmentIconNavFocusHighlightRawImage = navFocusHighlight.GetComponent<RawImage>();
+            equipmentIconNavFocusHighlightRawImage.texture = ModAssets.AssetBundle.LoadAsset<Texture2D>("NewNavHighlight");
+
+            HGButton button = equipmentBackground.GetComponent<HGButton>();
+            button.m_Colors.highlightedColor = Main.Helpers.ChangeColorWhileKeepingAlpha(button.m_Colors.highlightedColor, Color.white);
+            button.m_Colors.normalColor = Main.Helpers.ChangeColorWhileKeepingAlpha(button.m_Colors.normalColor, Color.white);
+            button.m_Colors.pressedColor = Main.Helpers.ChangeColorWhileKeepingAlpha(button.m_Colors.pressedColor, Color.white);
+            button.m_Colors.selectedColor = Main.Helpers.ChangeColorWhileKeepingAlpha(button.m_Colors.selectedColor, Color.white);
+            button.m_Colors.m_ColorMultiplier = 1.5f;
+        }
 
 
         internal static void SetSkillsAndEquipmentReminderTextStatus()